+++ test/fixtures/laravel-basic/app/Http/Controllers/Api/ProductStatisticController.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace App\Http\Controllers\Api;
+
+class ProductStatisticController
+{
+    public function index()
+    {
+        return response()->json([]);
+    }
+
+    public function store()
+    {
+        return response()->json([]);
+    }
+
+    public function show($productStatistic)
+    {
+        return response()->json([]);
+    }
+
+    public function update($productStatistic)
+    {
+        return response()->json([]);
+    }
+
+    public function destroy($productStatistic)
+    {
+        return response()->json([]);
+    }
+}